Aggregate Legend Query and Query Builder action header
