+++ .changeset/afraid-baboons-sit.md
@@ -1,0 +1,6 @@
+---
+'@finos/legend-application-query-bootstrap': patch
+'@finos/legend-application-query': patch
+'@finos/legend-query-builder': patch
+'@finos/legend-application': patch
+---
