+++ .changeset/orange-camels-swim.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-application-query': patch
+---
+
+Aggregate Legend Query and Query Builder action header
