+++ packages/legend-application-query-bootstrap/style/index.scss
@@ -938,6 +938,14 @@
       border: 0.1rem solid var(--color-legacylight-light-blue-250);
     }
 
+    .query-editor__header__action-combo {
+      background: var(--color-legacylight-light-blue-250);
+    }
+
+    .query-editor__header__action-combo svg {
+      color: var(--color-white);
+    }
+
     .query-builder__header {
       background: var(--color-white);
       border-bottom: none;
@@ -953,13 +961,14 @@
     }
 
     .query-builder__header__advanced-dropdown {
-      background: var(--color-legacylight-light-grey-100);
-      border-color: var(--color-light-grey-300);
+      background: var(--color-legacylight-light-blue-250);
+      color: var(--color-white);
     }
 
     .query-builder__header__advanced-dropdown__label,
+    .query-builder__header__advanced-dropdown svg,
     .query-builder__sub-header__custom-action__label {
-      color: var(--color-legacylight-dark-grey-400);
+      color: var(--color-white);
     }
 
     // ------------------------------------ Panel ---------------------------------------
