+++ packages/legend-query-builder/src/index.ts
@@ -29,8 +29,9 @@ export { QueryBuilderNavigationBlocker } from './components/QueryBuilderNavigati
 export { QueryBuilder } from './components/QueryBuilder.js';
 export { QUERY_BUILDER_COMPONENT_ELEMENT_ID } from './components/QueryBuilderComponentElement.js';
 export {
-  QueryBuilderState,
   type QuerySDLC,
+  QueryBuilderAdvancedButtonKey,
+  QueryBuilderState,
 } from './stores/QueryBuilderState.js';
 export {
   getTDSColumnDerivedProperyFromType,
