+++ packages/legend-query-builder/style/_query-builder.scss
@@ -42,6 +42,7 @@
     height: calc(100% - 3.2rem);
     width: 100%;
     padding: 0.3rem;
+    flex: 1;
   }
 
   &__header {
@@ -87,6 +88,7 @@
 
     justify-content: flex-end;
     flex: 1;
+    margin-right: 0.5rem;
 
     &__undo-redo {
       @include flexVCenter;
@@ -115,21 +117,20 @@
   &__header__advanced-dropdown {
     @include flexCenter;
 
-    height: 2.2rem;
+    height: 2.8rem;
     padding: 0 1rem;
     border-radius: 0.2rem;
-    margin-right: 0.5rem;
     font-weight: 500;
-    border: 0.1rem solid var(--color-dark-grey-300);
+    background: var(--color-blue-200);
 
     &__label {
       font-size: 1.2rem;
-      color: var(--color-dark-grey-500);
+      color: var(--color-light-grey-180);
     }
 
     &__icon {
       margin-left: 1rem;
-      color: var(--color-dark-grey-500);
+      color: var(--color-light-grey-180);
     }
 
     svg {
